Create LCAofBTSpaceOptimized.java
Naive Approach
The basic idea of this approach is to list down all the ancestors of the given nodes. And then, we will choose the common ancestor from the two lists which is farthest from the root of the tree.

 

Consider the binary tree as shown in the above figure, where we are trying to find the LCA of X and Y. Let us try to find the path from the root node to X and Y respectively and store the nodes present in the path in two separate lists pathToX and pathToY. Observe that the first few nodes(here nodes in yellow colour)  in the lists will be the same which are common ancestors of node X and Y.  Now, we need a common ancestor which is farthest from the root node. It is quite clear from the above image that the last common node in the lists will be the LCA.

 

Consider the following steps to describe the approach explained above:

Initialise two empty lists pathToX and pathToY to store the nodes in the path from the root node to ‘X’ and ‘Y’ respectively.
Perform a pre-order traversal to find the nodes in the path from the root node to X and store it in the pathToX such that the first element in the list is the root node.
Similarly, perform a pre-order traversal to find the nodes in the path from the root node to ‘Y’ and store it in the pathToY such that the first element in the list is the root node.
Now, start iterating from the front of the lists while the elements are equal.
The last equal element will be the LCA of X and Y.
Time Complexity
O(N), Where N is the number of nodes in the given binary tree.

 

Since we are doing two pre-order traversals and in the worst case (Skewed Trees), all nodes of the given tree can be present in the path when ‘X’ and ‘Y’ are very close to the leaf node of the binary tree. And each one can take O(N) time. So the overall time complexity will be O(N).

Space Complexity
O(N), Where N is the number of nodes in the given binary tree.

 

Since we are storing the paths in a list and in the worst case (Skewed Trees), all nodes of the given tree can be present in the list when ‘X’ will be the root node and ‘Y’ will be the leaf node of the binary tree. So the overall space complexity will be O(N).