Create widthOfaBT.java
Algorithm / Intuition
To determine the maximum width of a tree, an effective strategy would be to assign and identify indexes for the leftmost and rightmost nodes at each level. Using these indexes, we can calculate the width for each level by subtracting the index of the leftmost node from that of the rightmost node.


Start by assigning an index to the root node as 0. For each level, the left child gets an index equal to 2 * parent index, and the right child gets an index equal to 2 * parent index + 1. Using a level order traversal, we use the leftmost and rightmost nodes at each level and using their indices, get the width at that level. Keep track of the maximum width encountered during the traversal. Whenever a wider level is found, update the maximum width.


Algorithm:

Step 1:Initialize a variable `ans` to store the maximum width. If the root is null, return 0 as the width of an empty tree is zero.

Step 2: Create a queue to perform level-order traversal and each element of this queue would be a pair containing a node and its vertical index. Push the root node and its position (initially 0) into the queue.

Step 3: While the queue is not empty, perform the following steps:

Get the number of nodes at the current level (size).
Get the position of the front node in the current level which is the leftmost minimum index at that level.
Initialize variables first and last to store the first and last positions of nodes in the current level.
Step 4: Backtracking: For each node in the current level:

Calculate the current position relative to the minimum position in the level.
Get the current node (node) from the front of the queue.
If this is the first node in the level, update the first variable.
If this is the last node in the level, update the last variable.
Enqueue the left child of the current node with index: 2 x current index - 1.
Enqueue the right child of the current node with index: 2 x current index + 1.

Step 5: Update the maximum width (ans) by calculating the difference between the first and last positions, and adding 1.

Step 6: Repeat the level-order traversal until all levels are processed. The final value of `ans` represents the maximum width of the binary tree, return it.